Modifying operator_converters for keras to make onnx node naming correlate with keras model names:  * Activation  * AdvancedActivation  * BatchNorm  * Concat  * Conv  * Crop  * Dense  * Dot  * Flatten  * Permute  * Pool  * RepeatVector  * Reshape  * Upsample  * ZeroPad
In general if the operator expands to a single onnx node, it keeps the same name as the keras layer. If it expands to multiple onnx nodes, it follows: keras_name + '_' + operator_string (e.g. 'Dense_1_Add')
For the most part Tensor(Variable) names were attempted to be changed with a similar scheme